dom/network/tests/test_tcpsocket_enabled_no_perm.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test to ensure TCPSocket permission enabled and no tcp-socket perm does not allow open</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script type="application/javascript">

/** Test to ensure TCPSocket permission being turned on enables 
  navigator.mozTCPSocket, but mozTCPSocket.open does not work
  in content.
**/
SimpleTest.waitForExplicitFinish();
SpecialPowers.pushPrefEnv({"set": [['dom.mozTCPSocket.enabled', true]]}, runTest);
function runTest() {
  ok('mozTCPSocket' in navigator, "navigator.mozTCPSocket should be accessible if dom.mozTCPSocket.enabled  is true");

  try {
    navigator.mozTCPSocket.open('localhost', 80);
    throw new Error("Error: navigator.mozTCPSocket.open should raise for content that does not have the tcp-socket permission");
  } catch (e) {
      ok(true, "navigator.mozTCPSocket.open should raise for content that does not have the tcp-socket permission");
  }
  SimpleTest.finish();
}
</script>
</pre>
</body>
</html>
